not compiling: partially upgraded general_iter to u64

+++ src/util/general_iter.rs
@@ -36,7 +36,7 @@ macro_rules! gen_vector_iter_impl {
                     if Some(self.index) > self.back_index {
                         Err(std::num::NonZeroUsize::new(n).unwrap())
                     } else {
-                        Err(std::num::NonZeroUsize::new(n - (self.back_index.as_ref().unwrap_or(&usize::MAX).wrapping_sub(self.index).wrapping_add(1))).unwrap())
+                        Err(std::num::NonZeroUsize::new(n - (self.back_index.as_ref().unwrap_or(&u64::MAX).wrapping_sub(self.index).wrapping_add(1)) as usize).unwrap())
                     }
                 } else {
                     self.index += n;
@@ -62,10 +62,10 @@ macro_rules! gen_vector_iter_impl {
 
                 // since the cursors point to unconsumed items, we need to add 1
                 let remaining = *self.back_index.as_ref().unwrap() - self.index + 1;
-                if remaining < n {
-                    return Err(std::num::NonZeroUsize::new(n - remaining).unwrap());
+                if remaining < n as u64 {
+                    return Err(std::num::NonZeroUsize::new(n - remaining as usize).unwrap());
                 }
-                self.back_index = if self.back_index >= Some(n) { self.back_index.map(|b| b - n) } else { None };
+                self.back_index = if self.back_index >= Some(n as u64) { self.back_index.map(|b| b - n as u64) } else { None };
                 Ok(())
             }
 
@@ -126,7 +126,7 @@ macro_rules! gen_vector_iter_impl {
         impl $(<$life>)? std::iter::ExactSizeIterator for $name $(<$life>)? {
             fn len(&self) -> usize {
                 // intentionally overflowing calculations to avoid branches on empty iterator
-                (*self.back_index.as_ref().unwrap_or(&usize::MAX)).wrapping_sub(self.index).wrapping_add(1)
+                (*self.back_index.as_ref().unwrap_or(&u64::MAX)).wrapping_sub(self.index).wrapping_add(1) as usize
             }
         }
 
@@ -236,20 +236,20 @@ macro_rules! impl_vector_iterator {
         #[derive(Clone, Debug)]
         pub struct $own {
             vec: $type,
-            index: usize,
+            index: u64,
             // back index is none, iff it points to element -1 (i.e. element 0 has been consumed by
             // a call to next_back()). It can be Some(..) even if the iterator is empty
-            back_index: Option<usize>,
+            back_index: Option<u64>,
         }
 
         #[doc = concat!("A borrowing iterator for `", stringify!($type), "`.")]
         #[derive(Clone, Debug)]
         pub struct $bor<'a> {
             vec: &'a $type,
-            index: usize,
+            index: u64,
             // back index is none, iff it points to element -1 (i.e. element 0 has been consumed by
             // a call to next_back()). It can be Some(..) even if the iterator is empty
-            back_index: Option<usize>,
+            back_index: Option<u64>,
         }
 
         crate::util::gen_vector_iter_impl!($own, $type, $return_type, $get_unchecked, $get);
